+++ plugins/ui/src/js/src/elements/TextArea.tsx
@@ -1,14 +1,9 @@
-import React, { useCallback, useState } from 'react';
+import React from 'react';
 import {
   TextArea as DHCTextArea,
   TextAreaProps as DHCTextAreaProps,
 } from '@deephaven/components';
-import Log from '@deephaven/log';
-import { useDebouncedCallback, usePrevious } from '@deephaven/react-hooks';
-
-const log = Log.module('@deephaven/js-plugin-ui/TextArea');
-
-const VALUE_CHANGE_DEBOUNCE = 250;
+import useDebouncedOnChange from './hooks/useDebouncedOnChange';
 
 const EMPTY_FUNCTION = () => undefined;
 
@@ -24,49 +19,19 @@ export function TextArea(props: TextAreaProps): JSX.Element {
     ...otherProps
   } = props;
 
-  const [value, setValue] = useState(propValue ?? defaultValue);
-  const [pending, setPending] = useState(false);
-  const prevPropValue = usePrevious(propValue);
-
-  // Update local value to new propValue if the server sent a new propValue and no user changes have been queued
-  if (
-    propValue !== prevPropValue &&
-    propValue !== value &&
-    propValue !== undefined &&
-    !pending
-  ) {
-    setValue(propValue);
-  }
-
-  const propDebouncedOnChange = useCallback(
-    async (newValue: string) => {
-      try {
-        await propOnChange(newValue);
-      } catch (e) {
-        log.warn('Error returned from onChange', e);
-      }
-      setPending(false);
-    },
-    [propOnChange]
-  );
-
-  const debouncedOnChange = useDebouncedCallback(
-    propDebouncedOnChange,
-    VALUE_CHANGE_DEBOUNCE
-  );
-
-  const onChange = useCallback(
-    (newValue: string) => {
-      setPending(true);
-      debouncedOnChange(newValue);
-      setValue(newValue);
-    },
-    [debouncedOnChange]
+  const [value, onChange] = useDebouncedOnChange<string>(
+    propValue,
+    defaultValue,
+    propOnChange
   );
 
   return (
-    // eslint-disable-next-line react/jsx-props-no-spreading
-    <DHCTextArea value={value} onChange={onChange} {...otherProps} />
+    <DHCTextArea
+      value={value}
+      onChange={onChange}
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      {...otherProps}
+    />
   );
 }
